+++ tests/package.json
@@ -20,6 +20,7 @@
     "mocha": "^8.3.2",
     "mocha-steps": "^1.3.0",
     "rimraf": "^3.0.2",
+    "solc": "^0.8.3",
     "ts-node": "9.1",
     "typescript": "^3.9.6",
     "web3": "^1.3.5"
